+++ R/patients.R
@@ -1,4 +1,4 @@
-#' Converts a sample of patients in XLSX format into Unit Testing Definition in JSON format.
+#' Converts a sample of patients into Unit Testing Definition JSON file.
 #'
 #' @param filePath Path to the test patient data in Excel format. The Excel has sheets that represent tables from the OMOP-CDM, e.g. person, drug_exposure, condition_ocurrence, etc.
 #' @param testName A name of the test population in character.
@@ -9,11 +9,9 @@
 #'
 #' @importFrom readxl read_excel excel_sheets
 #' @importFrom jsonlite toJSON
-#' @importFrom usethis use_directory
-#' @importFrom fs path
-#' @importFrom usethis proj_path
 #' @importFrom checkmate assertDirectoryExists assertCharacter assertFileExists assert
 #' @importFrom glue glue
+#' @import cli
 #'
 #' @examples
 #' filePath <- system.file("extdata", "testPatientsRSV.xlsx", package = "TestGenerator")
@@ -24,41 +22,256 @@ readPatients <- function(filePath = NULL,
                          testName = "test",
                          outputPath = NULL,
                          cdmVersion = "5.3") {
+  checkmate::assertFileExists(filePath)
+  fileExtension <- tools::file_ext(filePath)
+  checkmate::assertTRUE(fileExtension %in% c("csv", "xlsx"))
+
+  if (fileExtension == "csv") {
+    readPatients.csv(filePath, testName, outputPath, cdmVersion)
+  } else {
+    readPatients.xl(filePath, testName, outputPath, cdmVersion)
+  }
+}
+
+#' Converts a sample of patients in XLSX format into Unit Testing Definition JSON file.
+#'
+#' @param filePath Path to the test patient data in Excel format. The Excel has sheets that represent tables from the OMOP-CDM, e.g. person, drug_exposure, condition_ocurrence, etc.
+#' @param testName A name of the test population in character.
+#' @param outputPath Path to write the test JSON files. If NULL, the files will be written at the project's testthat folder, i.e. tests/testthat/testCases.
+#' @param cdmVersion cdm version, default "5.3".
+#'
+#' @return A directory with the test JSON files with sample patients inside the project directory.
+#'
+#' @importFrom readxl read_excel excel_sheets
+#' @importFrom jsonlite toJSON
+#' @importFrom checkmate assertDirectoryExists assertCharacter assertFileExists assert
+#' @importFrom testthat test_path
+#' @importFrom glue glue
+#' @import cli
+#'
+#' @examples
+#' filePath <- system.file("extdata", "testPatientsRSV.xlsx", package = "TestGenerator")
+#' readPatients.xl(filePath = filePath, outputPath = tempdir())
+#'
+#' @export
+readPatients.xl <- function(filePath = NULL,
+                            testName = "test",
+                            outputPath = NULL,
+                            cdmVersion = "5.3") {
 
   checkmate::assertCharacter(filePath)
   checkmate::assertFileExists(filePath)
 
+  # Check columns
   expectedTables <- spec_cdm_field[[cdmVersion]] %>%
     dplyr::pull(cdmTableName)
 
   patientTables <- readxl::excel_sheets(filePath)
   checkmate::assert(all(patientTables %in% unique(expectedTables)))
 
-  listPatientTables <- lapply(patientTables,
+  cdmTables <- lapply(patientTables,
                               readxl::read_excel,
                               path = filePath)
 
-  names(listPatientTables) <- tolower(patientTables)
+  names(cdmTables) <- tolower(patientTables)
 
-  testCaseFile <- jsonlite::toJSON(listPatientTables,
+  # Convert to JSON
+  testCaseFile <- jsonlite::toJSON(cdmTables,
                                    dataframe = "rows",
                                    pretty = TRUE)
 
+  # Create testPath folder
+  testPath <- createOutputFolder(outputPath, testName)
+
+  # Write file
+  write(testCaseFile, file = testPath)
+  if (checkmate::checkFileExists(testPath)) {
+    cli::cli_alert_success(glue::glue("Unit Test Definition Created Successfully: '{testName}'"))
+  } else {
+    cli::cli_alert_danger("Unit Test Definition Creation Failed")
+    stop()
+  }
+}
+
+#' Converts a sample of patients in CSV format into a Unit Testing Definition JSON file.
+#'
+#' @param filePath Path to the test patient data in CSV format. Multiple CSV files representing tables tables from the OMOP-CDM must be provided, e.g. person.csv, drug_exposure.csv, condition_ocurrence.csv, etc.
+#' @param testName Name for the test population file in character.
+#' @param outputPath Path of the output file, if NULL, a folder will be created in the project folder inst/testCases.
+#' @param cdmVersion cdm version, default "5.3".
+#' @param reduceLargeIds Reduces the length of very long ids generally in int64 format, such as those found in the MIMIC-IV database.
+#'
+#' @return A JSON file with sample patients inside the project directory.
+#'
+#' @importFrom readr read_csv
+#' @importFrom jsonlite toJSON
+#' @importFrom checkmate assertDirectoryExists assertCharacter assertFileExists assert
+#' @importFrom glue glue
+#' @importFrom tools file_path_sans_ext
+#' @import cli
+#'
+#' @examples
+#' filePath <- system.file("extdata", "mimic_sample", package = "TestGenerator")
+#' readPatients.csv(filePath = filePath, outputPath = tempdir())
+#'
+#' @export
+readPatients.csv <- function(filePath = NULL,
+                             testName = "test",
+                             outputPath = NULL,
+                             cdmVersion = "5.3",
+                             reduceLargeIds = FALSE) {
+
+  checkmate::assertDirectoryExists(filePath)
+  checkmate::assertCharacter(cdmVersion)
+  checkmate::assertTRUE(cdmVersion %in% c("5.3", "5.4"))
+
+  # Check column
+  cdmTables <- fileColumnCheck(filePath, cdmVersion)
+
+  if (reduceLargeIds) {
+    cdmTables <- convertIds(cdmTables)
+  }
+
+  # Convert to JSON
+  testCaseFile <- jsonlite::toJSON(cdmTables,
+                                   dataframe = "rows",
+                                   pretty = TRUE)
+
+  # Create testPath folder
+  testPath <- createOutputFolder(outputPath, testName)
+
+  # Write file
+  write(testCaseFile, file = testPath)
+  if (checkmate::checkFileExists(testPath)) {
+    cli::cli_alert_success(glue::glue("Unit Test Definition Created Successfully: '{testName}'"))
+  } else {
+    cli::cli_alert_danger("Unit Test Definition Creation Failed")
+    stop()
+  }
+}
+
+fileColumnCheck <- function(filePath, cdmVersion) {
+  checkmate::assertDirectoryExists(filePath)
+  checkmate::assertCharacter(cdmVersion)
+  checkmate::assertTRUE(cdmVersion %in% c("5.3", "5.4"))
+  csvFiles <- list.files(filePath, pattern = ".csv", full.names = TRUE)
+  csvFilesNames <- list.files(filePath, pattern = ".csv")
+  checkmate::assertCharacter(csvFiles, any.missing = FALSE, min.len = 1)
+  checkmate::assertCharacter(csvFilesNames, any.missing = FALSE, min.len = 1)
+  currentTables <- spec_cdm_field[[cdmVersion]] %>%
+    dplyr::pull(cdmTableName) %>%
+    unique()
+  patientTables <- list()
+  report <- list()
+  for (i in 1:length(csvFiles)) {
+    tableName <- tools::file_path_sans_ext(csvFilesNames[i])
+    if (tableName %in% currentTables) {
+      cdmTable <- readr::read_csv(csvFiles[i], show_col_types = FALSE)
+      if (nrow(cdmTable) != 0) {
+        names(cdmTable) <- tolower(names(cdmTable))
+        currentCoulumns <- names(cdmTable)
+        expectedColumns <- spec_cdm_field[[cdmVersion]] %>%
+          dplyr::filter(cdmTableName == tableName) %>%
+          dplyr::pull(cdmFieldName)
+        expectedColumns <- gsub("\"", "", expectedColumns)
+        if (all(currentCoulumns %in% expectedColumns)) {
+          patientTables[[tableName]] <- cdmTable
+        } else {
+          cli::cli_alert_danger(glue::glue("'{tableName}' table columns do not match"))
+          stop()
+        }
+      } else {
+        report[["empty"]] <- append(report[["empty"]], glue::glue("{tableName}"))
+      }
+    }
+  }
+  if (!is.null(report[['empty']])) {
+    empty <- paste(report[['empty']], collapse = ", ")
+    cli::cli_alert_warning("Empty Tables Found:")
+    cli::cli_text(empty)
+  }
+  names(patientTables) <- tolower(names(patientTables))
+  return(patientTables)
+}
+
+convertIds <- function(cdmTables) {
+  report <- list()
+  for (tables in names(cdmTables)) {
+    # tables <- "vocabulary"
+    for (columns in names(cdmTables[[tables]])) {
+      # columns <- "vocabulary_concept_id"
+      if (columns %in% c("person_id",
+                         "care_site_id",
+                         "condition_era_id",
+                         "condition_occurrence_id",
+                         "device_exposure_id",
+                         "visit_occurrence_id",
+                         "dose_era_id",
+                         "drug_era_id",
+                         "drug_exposure_id",
+                         "fact_id_1",
+                         "fact_id_2",
+                         "measurement_id",
+                         "observation_id",
+                         "observation_period_id",
+                         "procedure_occurrence_id",
+                         "specimen_id",
+                         "visit_detail_id",
+                         "preceding_visit_detail_id",
+                         "preceding_visit_occurrence_id",
+                         "vocabulary_concept_id")) {
+
+        uniqueIdValues <- unique(cdmTables[[tables]][[columns]])
+        idValues <- abs(cdmTables[[tables]][[columns]]) %>%
+          abs() %>%
+          format(scientific = FALSE, trim = TRUE) %>%
+          substr(1, 9) %>%
+          as.numeric()
+
+        if (length(unique(idValues)) != length(unique(uniqueIdValues))) {
+          if(!tables %in% c("person_id", "visit_occurrence_id", "condition_occurrence_id")) {
+            cdmTables[[tables]][[columns]] <- seq(1, length(uniqueIdValues))
+            report[["notUnique"]] <- append(report[["notUnique"]], glue::glue("{tables}"))
+            # message(glue::glue("'{tables}' table with '{columns}' ids are not unique"))
+            # message(glue::glue("'{tables}' table filled out with sequence of numbers"))
+          } else {
+            cli::cli_alert_danger(glue::glue("'{tables}' table with '{columns}' ids are not unique and couldn't fill with num sequence"))
+            stop()
+          }
+        } else {
+          cdmTables[[tables]][[columns]] <- idValues
+          # report[["reduced"]] <- append(report[["reduced"]], glue::glue("{tables}"))
+          # message(glue::glue("'{tables}' table and '{columns}' ids reduced succesfully"))
+        }
+      }
+    }
+  }
+  if (!is.null(report[['notUnique']])) {
+    notUnique <- paste(report[['notUnique']], collapse = ", ")
+    cli::cli_alert_warning("Table with non unique ids and filled with num seq:")
+    cli::cli_text(notUnique)
+  }
+  cli::cli_alert_success("IDs successfully reduced")
+  # reduced <- paste(report[['reduced']], collapse = ", ")
+  # cli::cli_text(reduced)
+  return(cdmTables)
+}
+
+createOutputFolder <- function(outputPath, testName) {
   if (is.null(outputPath)) {
-    usethis::use_directory(fs::path("inst", "testCases"))
-    outputPath <- fs::path("inst", "testCases")
-    testPath <- paste0(outputPath, "/", testName, ".json")
+    testFolder <- testthat::test_path("testCases")
+    if (!dir.exists(testFolder)) {
+      dir.create(testFolder)
+      testPath <- paste0(testFolder, "/", testName, ".json")
+    } else {
+      testPath <- paste0(testFolder, "/", testName, ".json")
+    }
   } else {
     checkmate::assertCharacter(outputPath)
     checkmate::assertDirectoryExists(outputPath)
     testPath <- paste0(outputPath, "/", testName, ".json")
   }
-  write(testCaseFile, file = testPath)
-  if (checkmate::checkFileExists(testPath)) {
-    message(glue::glue("Unit Test Definition created successfully: {testName}"))
-  } else {
-    stop("Unit Test Definition creation failed")
-  }
+  return(testPath)
 }
 
 #' Pushes test population into a blank CDM.
@@ -68,8 +281,7 @@ readPatients <- function(filePath = NULL,
 #' @param cdmVersion cdm version, default "5.3".
 #'
 #' @return A CDM reference object with a sample population.
-#' @import dplyr
-#' @importFrom usethis proj_path
+#' @import dplyr cli
 #' @importFrom DBI dbConnect dbAppendTable dbDisconnect
 #' @importFrom duckdb duckdb
 #' @importFrom jsonlite fromJSON
@@ -88,14 +300,21 @@ patientsCDM <- function(pathJson = NULL,
                         cdmVersion = "5.3") {
 
   if (is.null(pathJson)) {
-    pathJson <- proj_path("inst", "testCases")
+    outputFolder <- testthat::test_path("testCases")
+    if (dir.exists(outputFolder)) {
+      pathJson <- outputFolder
+    } else {
+      cli::cli_alert_danger("testCases not found")
+      stop()
+    }
   }
 
   checkmate::assertClass(pathJson, "character")
   checkmate::assertDirectoryExists(pathJson)
 
   if (identical(list.files(pathJson), character(0))) {
-    stop("Directory empty. Provide Unit Test Definitions")
+    cli::cli_alert_danger("Directory empty. Provide Unit Test Definitions")
+    stop()
   }
 
   testFiles <- list.files(pathJson, pattern = ".json")
@@ -118,48 +337,26 @@ patientsCDM <- function(pathJson = NULL,
 
   # Check/Download vocabulary
 
-  if ("empty_cdm" %in% CDMConnector::example_datasets()) {
+  vocabPath <- file.path(Sys.getenv("EUNOMIA_DATA_FOLDER"),
+                         glue::glue("empty_cdm_{cdmVersion}.zip"))
 
-    vocabPath <- file.path(Sys.getenv("EUNOMIA_DATA_FOLDER"),
-                           glue::glue("empty_cdm_{cdmVersion}.zip"))
-
-    if (!file.exists(vocabPath)) {
-      CDMConnector::downloadEunomiaData(datasetName = "empty_cdm",
-                                        cdmVersion = cdmVersion,
-                                        pathToData = Sys.getenv("EUNOMIA_DATA_FOLDER"),
-                                        overwrite = TRUE)
-    }
-
-    conn <- DBI::dbConnect(duckdb::duckdb(CDMConnector::eunomia_dir("empty_cdm")))
-    cdm <- CDMConnector::cdmFromCon(con = conn,
-                                    cdmSchema = "main",
-                                    writeSchema = "main")
-
-  } else {
-
-    # Get allergies 10K dataset
-    vocabPath <- file.path(Sys.getenv("EUNOMIA_DATA_FOLDER"),
-                           glue::glue("synthea-allergies-10k_{cdmVersion}.zip"))
-
-    if (!file.exists(vocabPath)) {
-      CDMConnector::downloadEunomiaData(datasetName = "synthea-allergies-10k")
-    }
-
-    # Empty CDM
-    conn <- DBI::dbConnect(duckdb::duckdb(CDMConnector::eunomia_dir("synthea-allergies-10k")))
-    cdm <- CDMConnector::cdmFromCon(con = conn,
-                                    cdmSchema = "main",
-                                    writeSchema = "main")
-    cdm <- emptyCDM(conn = conn,
-                    cdm = cdm)
-
+  if (!file.exists(vocabPath)) {
+    CDMConnector::downloadEunomiaData(datasetName = "empty_cdm",
+                                      cdmVersion = cdmVersion,
+                                      pathToData = Sys.getenv("EUNOMIA_DATA_FOLDER"),
+                                      overwrite = TRUE)
   }
+
+  conn <- DBI::dbConnect(duckdb::duckdb(CDMConnector::eunomia_dir("empty_cdm")))
+  cdm <- CDMConnector::cdmFromCon(con = conn,
+                                  cdmSchema = "main",
+                                  writeSchema = "main")
 
   # Read the JSON file into R
   jsonData <- jsonlite::fromJSON(fileName)
   # Check for the expected columns in the CDM
   for (tableName in names(jsonData)) {
-    # tableName <- "visit_detail"
+    # tableName <- "vocabulary"
     currentCoulumns <- names(jsonData[[tableName]])
     expectedColumns <- spec_cdm_field[[cdmVersion]] %>%
       dplyr::filter(cdmTableName == tableName) %>%
@@ -169,11 +366,12 @@ patientsCDM <- function(pathJson = NULL,
   }
 
   # Convert the JSON data into a data frame and append it to the blank CDM
+
   for (tableName in names(jsonData)) {
-    # tableName <- "visit_occurrence"
+    # tableName <- "vocabulary"
     patientData <- as.data.frame(jsonData[[tableName]])
     DBI::dbAppendTable(conn, tableName, patientData)
   }
-  message(glue::glue("Patients pushed to blank CDM successfully"))
+  cli::cli_alert_success("Patients pushed to blank CDM successfully")
   return(cdm)
 }