+++ R/utilsCdm.R
@@ -31,4 +31,3 @@ emptyCDM <- function(conn, cdm) {
   }
   return(cdm)
 }
-